Fix namespace issues with base Presenter
I was getting weird errors with rspec because the base presenter was in
a different namespace to the children classes. Also, renamed to make it
clear is the superclass (Base).
